updates for crypto challenges and connectivity tests

+++ func_calling/setup.sh
@@ -4,8 +4,9 @@
 docker build -t ctfenv .
 
 # Build docker image for each challenge
-for d in chals/crypto/*; do
+for d in chals/{pwn,crypto}/*; do
     if [ -d "$d" ]; then
+        echo "Building $d"
         image_name=$(jq -r .container_image < "$d"/challenge.json)
         docker build -t "$image_name" "$d"
     fi
@@ -15,7 +16,8 @@ done
 docker network create ctfnet
 
 # Download and unpack Ghidra
-wget https://github.com/NationalSecurityAgency/ghidra/releases/download/Ghidra_11.0_build/ghidra_11.0_PUBLIC_20231222.zip
-unzip ghidra_11.0_PUBLIC_20231222.zip
-rm ghidra_11.0_PUBLIC_20231222.zip
-
+if [ ! -d ghidra_11.0_PUBLIC ]; then
+    wget https://github.com/NationalSecurityAgency/ghidra/releases/download/Ghidra_11.0_build/ghidra_11.0_PUBLIC_20231222.zip
+    unzip ghidra_11.0_PUBLIC_20231222.zip
+    rm ghidra_11.0_PUBLIC_20231222.zip
+fi
